requirements.txt
asgiref==3.7.2
Django==4.2.3
django-ckeditor==6.7.0
django-crispy-forms==1.11.2
django-js-asset==2.1.0
pytz==2021.1
sqlparse==0.4.1
typing_extensions==4.7.1
